requirements.txt
streamlit>=1.30.0
vanna>=0.2.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
openai>=1.5.0
chromadb>=0.4.18 